app/requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.1.3
fastapi==0.101.1
h11==0.14.0
idna==3.4
kombu==5.3.1
prompt-toolkit==3.0.39
pydantic==2.1.1
pydantic-settings==2.0.3
pydantic_core==2.4.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.6
